+++ bin/damt.dart
@@ -20,7 +20,7 @@ import 'package:dav/dav.dart';
 import 'package:damt/records.dart';
 
 // set values to be used with package: dav and for help output
-const String applicationVersion = "0.3.0";
+const String applicationVersion = "0.3.1";
 const String copyright = "Copyright © 2023 Simon Rowe <simon@wiremoons.com>";
 
 void main(List<String> arguments) async {
@@ -107,5 +107,6 @@ void main(List<String> arguments) async {
   stdout.writeln("Total acronyms:            ${damt.dbRecordCount}");
   stdout.writeln("Last acronym added:        ${damt.dbNewestAcronym}");
   stderr.writeln("\nValid options are:\n${parser.usage}");
+  damt.dbClose();
   exit(0);
 }